Crate rose_tree

rose_tree is a rose tree (aka multi-way tree) data structure library.

The most prominent type is RoseTree - a wrapper around petgraph (http://bluss.github.io/petulant-avenger-graphlibrary/doc/petgraph/index.html)’s [Graph] (http://bluss.github.io/petulant-avenger-graphlibrary/doc/petgraph/graph/struct.Graph.html) data structure, exposing a refined API targeted towards rose tree related functionality.

Re-exports

pub extern crate petgraph;

Structs

NodeIndex
Node identifier.

ParentRecursion
An iterator that yeilds an index to the parent of the current child before the setting the parent as the new current child. This occurs recursively until the root index is yeilded.

RoseTree
An indexable tree data structure with a variable and unbounded number of branches per node.

Siblings
An iterator yielding indices to the siblings of some child node.

WalkChildren
A “walker” object that can be used to step through the children of some parent node.

WalkSiblings
A “walker” object that can be used to step through the siblings of some child node.

Constants

ROOT
RoseTree’s API ensures that it always has a “root” node and that its index is always 0.

Type Definitions

Children
An iterator yielding indices to the children of some node.

PetGraph
The PetGraph to be used internally within the RoseTree for storing/managing nodes and edges.